Establish setting the boot time parameter
Boot time was absent because function which set up it were called after
sending cmdline to the kernel. To establish the correct sequence,
the parameter must be set in the function boot_prep_linux(). In this case,
the execution time of the functions boot_prep_linux() and
announce_and_cleanup() is lost. Thus, when setting the boot time parameter,
a constant (the time from boot_prep_linux() to boot_jump_linux()) is added
to the boot_time.

Jira-ID: ADM-2598
Change-Id: If2f75d52e1d5f22abc3bdf872348aaffa670e1db
Signed-off-by: a.lukianenko <a.lukianenko@globallogic.com>