r-car: Get Board ID from BL2
u-boot needs to know the board ID
for loading proper dts file. This
has to be done before I2C initialization
in u-boot. Since board ID is stored
in eeprom which is accessible by I2C,
the only way to get the board ID
is to pass it from BL2.

Jira-Id: ADM-2684

Change-Id: I583befa50399dda1acb0970fb07961e45dc922cb
Signed-off-by: Vitaliy Vasylskyy <vitaliy.vasylskyy@globallogic.com>